Split out separate tests from fitting_context_test
for readability
